Use CS Core on supported platforms for USB cameras
Can be more reliable then OpenCV, and handles disconnects properly
